+++ core/common/src/unsafe/UnsafeBufferOperations.kt
@@ -254,12 +254,11 @@ public object UnsafeBufferOperations {
     public inline fun writeToTail(
         buffer: Buffer,
         minimumCapacity: Int,
-        writeAction: (SegmentWriteContext, Segment) -> Int
+        writeAction: (context: SegmentWriteContext, tail: Segment) -> Int
     ): Int {
         contract {
             callsInPlace(writeAction, EXACTLY_ONCE)
         }
-
         val tail = buffer.writableSegment(minimumCapacity)
         val bytesWritten = writeAction(SegmentWriteContextImpl, tail)
 
@@ -296,15 +295,19 @@ public object UnsafeBufferOperations {
      * the [iterationAction].
      *
      * Both [iterationAction] arguments are valid only within [iterationAction] scope,
-     * it's an error to store and reuse it later.
+     * it is an error to store and reuse it later.
+     *
+     * For a full iteration over buffer's segments, see [forEachSegment].
      *
      * @param buffer a buffer to iterate over
      * @param iterationAction a callback to invoke with the head reference and an iteration context instance
      *
-     * @sample kotlinx.io.samples.unsafe.UnsafeReadWriteSamplesJvm.messageDigest
-     * @sample kotlinx.io.samples.unsafe.UnsafeBufferOperationsSamples.crc32Unsafe
+     * @sample kotlinx.io.samples.unsafe.UnsafeBufferOperationsSamples.crc32GetUnchecked
      */
-    public inline fun iterate(buffer: Buffer, iterationAction: (BufferIterationContext, Segment?) -> Unit) {
+    public inline fun iterate(
+        buffer: Buffer,
+        iterationAction: (context: BufferIterationContext, head: Segment?) -> Unit
+    ) {
         contract {
             callsInPlace(iterationAction, EXACTLY_ONCE)
         }
@@ -335,7 +338,7 @@ public object UnsafeBufferOperations {
      */
     public inline fun iterate(
         buffer: Buffer, offset: Long,
-        iterationAction: (BufferIterationContext, Segment?, Long) -> Unit
+        iterationAction: (context: BufferIterationContext, segment: Segment?, startOfTheSegmentOffset: Long) -> Unit
     ) {
         contract {
             callsInPlace(iterationAction, EXACTLY_ONCE)
@@ -348,6 +351,34 @@ public object UnsafeBufferOperations {
 
         buffer.seek(offset) { s, o ->
             iterationAction(BufferIterationContextImpl, s, o)
+        }
+    }
+
+    /**
+     * Iterates over [buffer] segments starting from the head.
+     *
+     * [action] is invoked with an instance of [SegmentReadContext]
+     * allowing to read and write in an unchecked manner from [buffer]'s segments
+     *
+     * It is considered an error to use a [SegmentReadContext] or a [Segment] instances outside the scope of
+     * the [action].
+     *
+     * Both [action] arguments are valid only within [action] scope, it is an error to store and reuse it later.
+     * The action might never be invoked if the given [buffer] is empty.
+     *
+     * @param buffer a buffer to iterate over
+     * @param action a callback to invoke with the head reference and an iteration context instance
+     * @sample kotlinx.io.samples.unsafe.UnsafeReadWriteSamplesJvm.messageDigest
+     * @sample kotlinx.io.samples.unsafe.UnsafeBufferOperationsSamples.crc32Unsafe
+     */
+    public inline fun forEachSegment(
+        buffer: Buffer,
+        action: (context: SegmentReadContext, segment: Segment) -> Unit
+    ) {
+        var curr: Segment? = buffer.head
+        while (curr != null) {
+            action(SegmentReadContextImpl, curr)
+            curr = curr.next
         }
     }
 }
@@ -393,7 +424,10 @@ public interface SegmentReadContext {
 @UnsafeIoApi
 @JvmSynthetic
 @OptIn(ExperimentalContracts::class)
-public inline fun SegmentReadContext.withData(segment: Segment, readAction: (ByteArray, Int, Int) -> Unit) {
+public inline fun SegmentReadContext.withData(
+    segment: Segment,
+    readAction: (bytes: ByteArray, startIndexInclusive: Int, endIndexExclusive: Int) -> Unit
+) {
     contract {
         callsInPlace(readAction, EXACTLY_ONCE)
     }
@@ -489,8 +523,7 @@ public interface BufferIterationContext : SegmentReadContext {
      *
      * @param segment a segment for which a successor needs to be found
      *
-     * @sample kotlinx.io.samples.unsafe.UnsafeReadWriteSamplesJvm.messageDigest
-     * @sample kotlinx.io.samples.unsafe.UnsafeBufferOperationsSamples.crc32Unsafe
+     * @sample kotlinx.io.samples.unsafe.UnsafeBufferOperationsSamples.crc32GetUnchecked
      */
     public fun next(segment: Segment): Segment?
 }